# api/openapi.yaml
openapi: 3.1.0
info:
  title: Kredit API
  version: 0.1.0
  description: Contract-first API for Kredit production v1.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /buyer/me/verification/refresh:
    parameters:
      - name: business_id
        in: query
        description: Select an owned business; defaults to the latest business.
        schema: {type: string, format: uuid}
    post:
      operationId: refreshBuyerVerification
      summary: Resume saved identity checks without recreating uncertain provider sessions
      description: Authenticated account and CSRF required. Pending verification does not permit accepting a sale.
      responses:
        '200': {description: Current buyer portal and verification status}
        '503': {description: Details remain saved; verification is pending or needs provider review}
...
  /auth/otp/challenges:
    post:
      security: []
      operationId: requestOTP
      summary: Request a passwordless OTP challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTPRequest'
      responses:
        '202':
          description: Challenge accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OTPChallengeResponse'
        '429':
          description: Rate limited or resend cooldown active.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /organizations/{organizationID}/buyer-invitations:
    parameters:
      - $ref: '#/components/parameters/OrganizationID'
    post:
      parameters:
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      operationId: createBuyerInvitation
      summary: Create a secure buyer onboarding invitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyerInvitationRequest'
      responses:
        '202':
          description: Buyer invitation created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyerInvitationCreatedResponse'
  /buyer-invitations/{token}:
    parameters:
      - $ref: '#/components/parameters/InvitationToken'
    get:
      security: []
      operationId: previewBuyerInvitation
      summary: Preview a secure buyer invitation
      responses:
        '200':
          description: Safe invitation preview.
        '404':
          description: Invalid or expired invitation.
  /buyer-invitations/{token}/otp:
    parameters:
      - $ref: '#/components/parameters/InvitationToken'
    post:
      security: []
      operationId: requestBuyerInvitationOTP
      summary: Request OTP for the intended buyer contact
      responses:
        '202':
          description: Buyer verification challenge created.
  /buyer-invitations/{token}/accept:
    parameters:
      - $ref: '#/components/parameters/InvitationToken'
    post:
      parameters:
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      security: []
      operationId: acceptBuyerInvitation
      summary: Verify the buyer and create the buyer portal account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyerInvitationAcceptRequest'
      responses:
        '201':
          description: Buyer account and consent evidence saved; verification may still be pending.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyerPortalResponse'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: kredit_session
  parameters:
    RequiredIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    OrganizationID:
      name: organizationID
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
    InvitationToken:
      name: token
      in: path
      required: true
      schema:
        type: string
        minLength: 64
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        request_id:
          type: string
    OTPRequest:
      type: object
      required: [identifier, channel]
      properties:
        identifier:
          type: string
        channel:
          type: string
          enum: [phone, email]
        purpose:
          type: string
          default: login
    OTPChallengeResponse:
      type: object
      required: [challenge_id, expires_at, channel, message]
      properties:
        challenge_id:
          type: string
        expires_at:
          type: string
          format: date-time
        channel:
          type: string
        message:
          type: string
        development_code:
          type: string
          description: Development-only value; never returned in production.
...
    BuyerInvitationRequest:
      type: object
      required: [target, target_type, legal_name, business_type, business_address, industry]
      properties:
        target:
          type: string
        target_type:
          type: string
          enum: [phone, email]
        legal_name:
          type: string
        trading_name:
          type: string
        business_type:
          type: string
        business_address:
          type: string
        industry:
          type: string
    BuyerInvitationCreatedResponse:
      type: object
      required: [invitation, invitation_url]
      properties:
        invitation:
          $ref: '#/components/schemas/BuyerInvitation'
        invitation_url:
          type: string
          format: uri
    BuyerInvitation:
      type: object
      required: [id, organization_id, target_type, proposed_legal_name, proposed_business_type, proposed_address, proposed_industry, status, expires_at, created_at]
      properties:
        id:
          type: string
        organization_id:
          type: string
        target_type:
          type: string
        proposed_legal_name:
          type: string
        proposed_trading_name:
          type: string
        proposed_business_type:
          type: string
        proposed_address:
          type: string
        proposed_industry:
          type: string
        status:
          type: string
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    BuyerInvitationAcceptRequest:
      type: object
      required: [challenge_id, code, full_name, consents_accepted, terms_version, privacy_version, identity_notice_version]
      properties:
        consents_accepted: {type: boolean, const: true}
        terms_version: {type: string}
        privacy_version: {type: string}
        identity_notice_version: {type: string}
        challenge_id:
          type: string
        code:
          type: string
          minLength: 6
          maxLength: 6
        device_label:
          type: string
        full_name:
          type: string
        legal_name:
          type: string
        trading_name:
          type: string
        business_type:
          type: string
        business_address:
          type: string
        industry:
          type: string
    BuyerPerson:
      type: object
      required: [id, user_id, full_name, status, created_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        full_name:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
    BuyerBusiness:
      type: object
      required: [id, owner_user_id, legal_name, business_type, business_address, industry, status, created_at]
      properties:
        id:
          type: string
        owner_user_id:
          type: string
        legal_name:
          type: string
        trading_name:
          type: string
        business_type:
          type: string
        business_address:
          type: string
        industry:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
    BuyerRepresentative:
      type: object
      required: [id, business_id, person_id, role_title, authority_type, authority_status, created_at]
      properties:
        id:
          type: string
        business_id:
          type: string
        person_id:
          type: string
        role_title:
          type: string
        authority_type:
          type: string
        authority_status:
          type: string
        created_at:
          type: string
          format: date-time
    VerificationCase:
      type: object
      required: [id, subject_type, subject_id, provider, provider_reference, verification_level, state, started_at]
      properties:
        id:
          type: string
        subject_type:
          type: string
        subject_id:
          type: string
        provider:
          type: string
        provider_reference:
          type: string
        verification_level:
          type: integer
        state:
          type: string
        safe_result:
          type: object
          additionalProperties:
            type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    BuyerConsent:
      type: object
      required: [id, user_id, consent_type, version, accepted_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        consent_type:
          type: string
        version:
          type: string
        accepted_at:
          type: string
          format: date-time
    BuyerPortal:
      type: object
      required: [person, business, representative, verification_cases, consents, bank_accounts]
      properties:
        person:
          $ref: '#/components/schemas/BuyerPerson'
        business:
          $ref: '#/components/schemas/BuyerBusiness'
        representative:
          $ref: '#/components/schemas/BuyerRepresentative'
        verification_cases:
          type: array
          items:
            $ref: '#/components/schemas/VerificationCase'
        consents:
          type: array
          items:
            $ref: '#/components/schemas/BuyerConsent'
        bank_accounts:
          type: array
          items:
            type: object
    BuyerPortalResponse:
      type: object
      required: [portal]
      properties:
        verification_current:
          type: boolean
          description: All three identity checks are verified at the required level and have not expired.
        portal:
          $ref: '#/components/schemas/BuyerPortal'